pyproject.toml
[project]
name = "letters"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "shaun", email = "shaunhegarty@users.noreply.github.com"},
]
dependencies = [
    "fastapi==0.104.1",
    "psycopg2-binary==2.9.3",
    "uvicorn>=0.24.0.post1",
    "requests==2.31.0",
    "SQLAlchemy==2.0.23",
    "sqlmodel==0.0.14",
    "pydantic==2.5.2",
    "pydantic-core==2.14.5",
    "SQLAlchemy-Utils==0.41.1",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"

[tool.pdm.dev-dependencies]
typing = [
    "annotated-types==0.6.0",
]
test = [
    "pytest>=7.4.3",
    "httpx>=0.25.2",
]

[tool.pytest.ini_options]
pythonpath = "src"